# openapi.yaml
openapi: 3.1.0
info:
  title: Prelude API
  version: 0.0.0
  description: Prelude music server REST API specification
  license:
    name: GNU General Public License v3.0
    url: https://github.com/prelude-music/server/blob/main/LICENSE
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        description: The base URL of your Prelude server
        default: "http://localhost:9847"
...
paths:
...
  /tokens:
    get:
      tags: [Tokens]
      summary: List Tokens
      security:
        - basic: [tokens:read:self]
        - bearer: [tokens:read:self]
        - basic: [tokens:read:all]
        - bearer: [tokens:read:all]
      operationId: listTokens
      parameters: [
        {
          in: query,
          description: Number of resources to return per page,
          name: limit,
          schema: {
            type: integer,
            default: 100,
            minimum: 1
          }
        },
        {
          in: query,
          description: Page number to return,
          name: page,
          schema: {
            type: integer,
            default: 1,
            minimum: 1
          }
        },
        {
          in: query,
          description: Get tokens for a specific user. Incompatible with the `note` parameter.,
          name: user,
          schema: {
            $ref: "#/components/schemas/User/properties/id"
          }
        },
        {
          in: query,
          description: "Get ALL tokens. By default, only the tokens of the authenticated user are returned.
            Incompatible with the `user` parameter.",
          name: all,
          required: false,
          schema: {
            type: string,
            examples: [""]
          }
        }
      ]
      responses:
        200:
          description: A page of tokens
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      resources:
                        type: array
                        items:
                          $ref: "#/components/schemas/Token"
        401:
          $ref: "#/components/responses/Unauthorised"
        403:
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Tokens]
      summary: Create Token
      security:
        - basic: [tokens:write:self]
        - bearer: [tokens:write:self]
        - basic: [tokens:write:all]
        - bearer: [tokens:write:all]
      operationId: createToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [user, expires, scopes, note]
              properties:
                user:
                  description: The ID of the user to create this token for. You can only use this if you have the `tokens:write:all` scope.
                  allOf:
                    - $ref: "#/components/schemas/User/properties/id"
                expires:
                  $ref: "#/components/schemas/Token/properties/expires"
                scopes:
                  $ref: "#/components/schemas/Token/properties/scopes"
                note:
                  $ref: "#/components/schemas/Token/properties/note"
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [user, expires, scopes, note]
              properties:
                user:
                  description: The ID of the user to create this token for. You can only use this if you have the `tokens:write:all` scope.
                  allOf:
                    - $ref: "#/components/schemas/User/properties/id"
                expires:
                  $ref: "#/components/schemas/Token/properties/expires"
                scopes:
                  $ref: "#/components/schemas/Token/properties/scopes"
                note:
                  $ref: "#/components/schemas/Token/properties/note"
      responses:
        201:
          description: The created token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Token"
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: The token secret key used for authenticating requests. This is never available again.
                        examples: [ED80gmnuZumZ5N2kA0exhN2FVjpdBdk2jVze]
        401:
          $ref: "#/components/responses/Unauthorised"
        403:
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Tokens]
      summary: Delete All Tokens
      security:
        - basic: [tokens:write:self]
        - bearer: [tokens:write:self]
        - basic: [tokens:write:all]
        - bearer: [tokens:write:all]
      operationId: deleteAllTokens
      responses:
        204:
          description: All tokens deleted successfully
        401:
          $ref: "#/components/responses/Unauthorised"
        403:
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    basic:
      type: http
      scheme: Basic
      description: Authenticate using username and password. All scopes granted on your user account will apply. Please use HTTPS.
    bearer:
      type: http
      scheme: Bearer
      description: Authenticate using an API Token.
  responses:
    Unauthorised:
      description: |-
        Authorisation is missing or invalid. Possible causes for authorisation to be invalid are
          - the username or password is incorrect
          - the token secret is incorrect
          - the token secret is correct, but the token has expired
          - the username and password, or the token secret, are/is correct, but the user is disabled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The scopes of the user or token used to authenticate this request lacked the required scope for this endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Token:
      type: object
      properties:
        id:
          type: string
          title: Token ID
          examples: [7e10e4ff-be3f-46d0-9865-941c3d6b62c8]
        user:
          $ref: "#/components/schemas/User/properties/id"
        expires:
          description: The date in Unix time when the token expires. If null, the token does not expire. After the token expires, the server will consider it invalid.
          type:
            - integer
            - "null"
          format: unix-time
          examples: [1893448800]
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/Scope"
          title: Token scopes
        note:
          type: string
          description: A user-specified note for the token (e.g. description of what this token is for).
          maxLength: 128
    Scope:
      description: Controls access to API resources
      type: string
      enum: [library:read, library:write, tokens:read:self, tokens:write:self, tokens:read:all, tokens:write:all, users:read, users:write]
    Page:
      description: A page is a partial list of items from a larger collection
      type: object
      properties:
        page:
          description: The current page number
          type: integer
          examples: [1]
          minimum: 1
        limit:
          description: The maximum number of items per page
          examples: [100]
          minimum: 0
        total:
          description: The total number of items in the entire collection (not just the current page)
          examples: [91]
          minimum: 0
        href:
          description: The path and query parameters of the current page
          type: string
          examples: [/path?and=query]
        previous:
          description: The path and query parameters for the previous page, unless this is the first page
          type:
            - string
            - "null"
          examples: [null]
        next:
          description: The path and query parameters for the next page, unless this is the last page
          type:
            - string
            - "null"
          examples: [null]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              examples: [A user-friendly error message.]
            fields:
              type: object
...